// functions/portal/src/types.d.ts
export type DatabaseDialect = (typeof DATABASE_DIALECTS)[number]
// ...
export interface IDatasetDto {
  id: string
  type?: string
  tokenDatasetCode: string
  tenantId: string
  databaseCode: string
  schemaName: string
  vocabSchemaName: string
  dataModel: string
  paConfigId: string
  detail: IDatasetDetailBaseDto
  dashboards: IDatasetDashboardBaseDto[]
  attributes: IDatasetAttribute[]
  tags: string[]
  fhir_project_id?: string
}

export interface IDatasetSnapshotDto {
  id: string
  sourceDatasetId: string
  newDatasetName: string
  schemaName: string
  timestamp: Date
}

interface IDatasetDetailBaseDto {
  name: string
  summary: string
  description: string
  showRequestAccess: boolean
}

// TODO: Remove when we switch from study to dataset entirely
export interface IStudyDetailDto extends IDatasetDetailBaseDto {
  studyId: string
}

interface IDatasetDashboardBaseDto {
  name: string
  url: string
  basePath: string
}

interface IDatasetTag {
  id: number
  name: string
}

interface IDatasetAttribute {
  attributeId: string
  value: string
}

export interface IDatasetAttributeDto extends IDatasetAttribute {
  studyId: string
}

export type VisibilityStatus = (typeof VISIBILITY_STATUS)[number]

export type CdmSchemaOption = (typeof CDM_SCHEMA_OPTIONS)[number]